major refactor of rt.py into separate files
