+++ rayflare/ray_tracing/__init__.py
@@ -1,1 +1,2 @@
-from .rt import RT, rt_structure
+from .rt_matrix import RT
+from .rt_structure import rt_structure
