+++ rayflare/textures/define_textures.py
@@ -5,7 +5,7 @@
 #
 # Contact: p.pearce@unsw.edu.au
 
-from rayflare.ray_tracing.rt import RTSurface
+from rayflare.ray_tracing.rt_common import RTSurface
 import numpy as np
